Updated to 21.0.4
-Added option Claimed_Block_Limit to Damage_Detector. This is the
max damage a player can output to a block inside a claimed space
-Added console command to remove hardcore status from a player
while set to optional mode
-Changed target fps to 60
-Changed Damage_Detector options Entity_Damage_Limit to
Entity_Limit, Block_Damage_Limit to Block_Limit and
Player_Damage_Limit to Player_Limit
-Changed Output_Log to Output_Log_Blocker. Enabling creates
OutputBlocker.xml. Enter a phrase or sentence you want to block
from seeing in the output log. Helpful for spam and clutter
-Changed Bank deposit process to avoid lag issues
-Changed xml update process to use a timer delay when upgrading
the file. Should avoid the file being reported in use
-Changed Home tool to remove saved points if no claim is found when
attempting to use them
-Changed Damage_Detector logs to be written to file every five
seconds to avoid lag caused by rapid fire weapons and tools
-Updated Damage_Detector for blocks to match a21 process and to
detect irregular block damage inside a claimed space
-Updated Damage_Detector to skip players using items marked as dev
in the items.xml
-Updated memory allocations to reduce garbage collection
-Updated Block_Pickup detection to utilize more applicable data
-Updated phrase file entry Homes5, DamageDetector2
-Updated various tools for improved runtimes and reduced memory use
-Fixed bank to bag transfer not functioning as intended when over
the stack limit
-Fixed Player_Logs delay running in minutes instead of seconds
-Fixed Chat_Color utilizing multiple colors
-Fixed Vault_Box tool failing to save data on some servers
-Fixed Auto_Backup crashing the server when attempting to save
a file that is in use
-Fixed POI_Protection not functioning as intended
-Fixed console command 'st-Wallet All' not functioning as intended
-Fixed xml upgrade duplicating existing entries
-Removed non ommitted empty value examples from the xml files
-Removed phrase entry DamageDetector3

Known issues:
-Vault_Box does not work with all modded servers.
Might be a mod load order issue
-Shop panel does not recognize two players using the same ip
-RIO is not functioning as intended
-Spawn entity system not working as intended
-Some console commands are out of date

+++ ServerTools/src/Tools/Travel/Travel.cs
@@ -120,12 +120,15 @@ namespace ServerTools
                 }
                 else
                 {
-                    File.Delete(FilePath);
-                    if (childNodes != null && childNodes[0] != null)
+                    XmlNodeList nodeList = xmlDoc.DocumentElement.ChildNodes;
+                    if (nodeList != null)
                     {
-                        UpgradeXml(childNodes);
+                        File.Delete(FilePath);
+                        Timers.UpgradeTravelXml(nodeList);
+                        //UpgradeXml(childNodes);
                         return;
                     }
+                    File.Delete(FilePath);
                     UpdateXml();
                     return;
                 }
@@ -378,7 +381,7 @@ namespace ServerTools
             }
         }
 
-        private static void UpgradeXml(XmlNodeList nodeList)
+        public static void UpgradeXml(XmlNodeList nodeList)
         {
             try
             {
@@ -391,14 +394,16 @@ namespace ServerTools
                     sw.WriteLine("    <!-- <Location Name=\"zone2\" Corner1=\"-1,100,-1\" Corner2=\"-10,100,-10\" Destination=\"100,-1,100\" /> -->");
                     for (int i = 0; i < nodeList.Count; i++)
                     {
-                        if (!nodeList[i].OuterXml.Contains("<!-- <Location Name=\"Zone1\"") &&
+                        if (nodeList[i].NodeType == XmlNodeType.Comment)
+                        {
+                            if (!nodeList[i].OuterXml.Contains("<!-- <Location Name=\"Zone1\"") &&
                             !nodeList[i].OuterXml.Contains("<!-- <Location Name=\"Zone2\"") && !nodeList[i].OuterXml.Contains("<Location Name=\"\"") &&
                             !nodeList[i].OuterXml.Contains("<!-- <Version"))
-                        {
-                            sw.WriteLine(nodeList[i].OuterXml);
+                            {
+                                sw.WriteLine(nodeList[i].OuterXml);
+                            }
                         }
                     }
-                    sw.WriteLine("    <Location Name=\"\" Corner1=\"\" Corner2=\"\" Destination=\"\" />");
                     for (int i = 0; i < nodeList.Count; i++)
                     {
                         if (nodeList[i].NodeType != XmlNodeType.Comment)
